Closes #3 (#4)
This mapping set is a manually curated alignment of the LTER Controlled Vocabulary and Environment Ontology (ENVO) and was created using the Simple Standard for Sharing Ontology Mappings (SSSOM; https://mapping-commons.github.io/sssom/home/). The motivation of this mapping is to support semantic search of datasets in the Environmental Data Initiative (EDI) Data Repository. Even though the end-user will not see the mapping, rather only the results of it, a high degree of precision was attempted so as not to preclude the use in other contexts.

Note, the ontology mapping technique was a blend of element-level definitions and structure-level similarities, the latter being heavily used for the LTER Controlled Vocabulary for which the vast majority of concepts lacked element-level definitions. It should also be noted that an overly liberal assignment of skos:exactMatch to undefined concepts was used in this mapping set. A more cautious approach might be to assign skos:closeMatch in these situations.